+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
     * New task to notify professionals about Assignments with late submissions (rudaporto).
     * New task to notify professionals before shooting (rudaporto).
     * Card #462: two new _custom_filter for assignment endpoint: late_first_submission and late_re_submission (rudaporto).
+    * Card #462: two new _custom_filter for order endpoint: late_first_submission and late_re_submission (rudaporto).
 
 2.1.27 (2017-07-12)
 -------------------